[BUGFIX] Use realpath to clear symbolic link
The bug occurs when there is a /home/user1 and a /home2/user1. In cPanel,
it is possible for /home2 to be a symbolic link to /home. When install
path and access path are different, ABSPATH uses /home and get_home_path()
uses /home2. This causes the old code to try to write to //.htaccess.

realpath() solves the symbolic link issue, and trailingslashit is used
to make it consistent with the old code. realpath() returns a path
without a trailing slash, so if during refactoring, we want to remove
that call, need to be careful with the rest of the function.